+++ alacritty/src/config/bindings.rs
@@ -86,15 +86,20 @@ impl<T: Eq> Binding<T> {
         mode: TermMode,
         mods: ModifiersState,
         input: &T,
-        relaxed: bool,
     ) -> bool {
+        // Match `MOUSE_MODE` as any one mouse mode, since multiple can't be active at once
+        let mut binding_mode = self.mode;
+        if binding_mode.contains(TermMode::MOUSE_MODE) && mode.intersects(TermMode::MOUSE_MODE) {
+            binding_mode &= mode | !TermMode::MOUSE_MODE;
+        }
+
         // Check input first since bindings are stored in one big list. This is
         // the most likely item to fail so prioritizing it here allows more
         // checks to be short circuited.
         self.trigger == *input
-            && mode.contains(self.mode)
+            && mode.contains(binding_mode)
             && !mode.intersects(self.notmode)
-            && (self.mods == mods || (relaxed && self.mods.relaxed_eq(mods)))
+            && (self.mods == mods)
     }
 
     #[inline]
@@ -207,18 +212,6 @@ impl From<&'static str> for Action {
     }
 }
 
-pub trait RelaxedEq<T: ?Sized = Self> {
-    fn relaxed_eq(&self, other: T) -> bool;
-}
-
-impl RelaxedEq for ModifiersState {
-    // Make sure that modifiers in the config are always present,
-    // but ignore surplus modifiers.
-    fn relaxed_eq(&self, other: Self) -> bool {
-        !*self | other == ModifiersState::all()
-    }
-}
-
 macro_rules! bindings {
     (
         KeyBinding;
@@ -279,7 +272,8 @@ macro_rules! bindings {
 pub fn default_mouse_bindings() -> Vec<MouseBinding> {
     bindings!(
         MouseBinding;
-        MouseButton::Middle; Action::PasteSelection;
+        MouseButton::Middle, ModifiersState::SHIFT, +TermMode::MOUSE_MODE; Action::PasteSelection;
+        MouseButton::Middle, ~TermMode::MOUSE_MODE; Action::PasteSelection;
     )
 }
 
@@ -507,7 +501,7 @@ impl<'a> Deserialize<'a> for ModeWrapper {
             type Value = ModeWrapper;
 
             fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                f.write_str("Combination of AppCursor | AppKeypad, possibly with negation (~)")
+                f.write_str("Combination of Mouse | AppCursor | AppKeypad | Alt, possibly with negation (~)")
             }
 
             fn visit_str<E>(self, value: &str) -> Result<ModeWrapper, E>
@@ -518,12 +512,14 @@ impl<'a> Deserialize<'a> for ModeWrapper {
 
                 for modifier in value.split('|') {
                     match modifier.trim().to_lowercase().as_str() {
+                        "mouse" => res.mode |= TermMode::MOUSE_REPORT_CLICK,
+                        "~mouse" => res.not_mode |= TermMode::MOUSE_REPORT_CLICK,
                         "appcursor" => res.mode |= TermMode::APP_CURSOR,
                         "~appcursor" => res.not_mode |= TermMode::APP_CURSOR,
                         "appkeypad" => res.mode |= TermMode::APP_KEYPAD,
                         "~appkeypad" => res.not_mode |= TermMode::APP_KEYPAD,
-                        "~alt" => res.not_mode |= TermMode::ALT_SCREEN,
                         "alt" => res.mode |= TermMode::ALT_SCREEN,
+                        "~alt" => res.not_mode |= TermMode::ALT_SCREEN,
                         _ => error!(target: LOG_TARGET_CONFIG, "Unknown mode {:?}", modifier),
                     }
                 }
@@ -1040,8 +1036,8 @@ mod test {
         let mods = binding.mods;
         let mode = binding.mode;
 
-        assert!(binding.is_triggered_by(mode, mods, &13, true));
-        assert!(!binding.is_triggered_by(mode, mods, &32, true));
+        assert!(binding.is_triggered_by(mode, mods, &13));
+        assert!(!binding.is_triggered_by(mode, mods, &32));
     }
 
     #[test]
@@ -1055,14 +1051,9 @@ mod test {
         let t = binding.trigger;
         let mode = binding.mode;
 
-        assert!(binding.is_triggered_by(mode, binding.mods, &t, true));
-        assert!(binding.is_triggered_by(mode, binding.mods, &t, false));
-
-        assert!(binding.is_triggered_by(mode, superset_mods, &t, true));
-        assert!(!binding.is_triggered_by(mode, superset_mods, &t, false));
-
-        assert!(!binding.is_triggered_by(mode, subset_mods, &t, true));
-        assert!(!binding.is_triggered_by(mode, subset_mods, &t, false));
+        assert!(binding.is_triggered_by(mode, binding.mods, &t));
+        assert!(!binding.is_triggered_by(mode, superset_mods, &t));
+        assert!(!binding.is_triggered_by(mode, subset_mods, &t));
     }
 
     #[test]
@@ -1073,9 +1064,9 @@ mod test {
         let t = binding.trigger;
         let mods = binding.mods;
 
-        assert!(!binding.is_triggered_by(TermMode::INSERT, mods, &t, true));
-        assert!(binding.is_triggered_by(TermMode::ALT_SCREEN, mods, &t, true));
-        assert!(binding.is_triggered_by(TermMode::ALT_SCREEN | TermMode::INSERT, mods, &t, true));
+        assert!(!binding.is_triggered_by(TermMode::INSERT, mods, &t));
+        assert!(binding.is_triggered_by(TermMode::ALT_SCREEN, mods, &t));
+        assert!(binding.is_triggered_by(TermMode::ALT_SCREEN | TermMode::INSERT, mods, &t));
     }
 
     #[test]
@@ -1086,8 +1077,8 @@ mod test {
         let t = binding.trigger;
         let mods = binding.mods;
 
-        assert!(binding.is_triggered_by(TermMode::INSERT, mods, &t, true));
-        assert!(!binding.is_triggered_by(TermMode::ALT_SCREEN, mods, &t, true));
-        assert!(!binding.is_triggered_by(TermMode::ALT_SCREEN | TermMode::INSERT, mods, &t, true));
+        assert!(binding.is_triggered_by(TermMode::INSERT, mods, &t));
+        assert!(!binding.is_triggered_by(TermMode::ALT_SCREEN, mods, &t));
+        assert!(!binding.is_triggered_by(TermMode::ALT_SCREEN | TermMode::INSERT, mods, &t));
     }
 }